==> code/01_wine_webscraping.ipynb <==
import pandas as pd

# NLP-specific:
import requests
from bs4 import BeautifulSoup

---

## Scraping the Red Wine product entries from the URL and creating a DataFrame

- The url __[Wine.com](https://www.wine.com/list/wine/red-wine/7155-124?sortBy=mostInteresting)__ was webscrapted to gather information about a selection of popular red wine products.
- This information included:
> 1. Product ID
> 2. Wine Type
> 3. Wine Name
> 4. Wine Origin
> 5. Average Rating
> 6. Numerical Rating
> 7. Current Price
> 8. Prediscounted Price
> 9. Nominal Discount Value
> 10. Percent Savings

# Define the URLs
urls = [
    'https://www.wine.com/list/wine/red-wine/7155-124?sortBy=mostInteresting',
    'https://www.wine.com/list/wine/red-wine/7155-124?sortBy=mostPopular',
    'https://www.wine.com/list/wine/red-wine/7155-124?sortBy=topRated',
    'https://www.wine.com/list/wine/red-wine/7155-124?sortBy=userRatingDesc',
    'https://www.wine.com/list/wine/red-wine/7155-124?sortBy=priceLowToHigh',
    'https://www.wine.com/list/wine/red-wine/7155-124?sortBy=priceHighToLow',
    'https://www.wine.com/list/wine/red-wine/7155-124?sortBy=savings',
    'https://www.wine.com/list/wine/red-wine/7155-124?sortBy=justIn',
    'https://www.wine.com/list/wine/red-wine/7155-124?sortBy=oldToNew',
    'https://www.wine.com/list/wine/red-wine/7155-124?sortBy=newToOld',
    'https://www.wine.com/list/wine/red-wine/7155-124?sortBy=wineryAToZ',
    'https://www.wine.com/list/wine/red-wine/7155-124?sortBy=wineryZToA'
]

def scrape_wine_data(url):
    req = requests.get(url)
    soup = BeautifulSoup(req.content, 'html.parser')
    results = []

    # Extracting the number of elements available for each property
    num_elements = len(soup.body.find_all('meta', class_='schema_productID'))

    for idx in range(num_elements):
        new_entry = {}
        
        # Define the elements and their corresponding class names
        elements_info = [
            ('product_id', 'meta', 'schema_productID'),
            ('wine_type', 'span', 'listGridItemOrigin_varietal'),
            ('wine_name', 'span', 'listGridItemInfo_name'),
            ('wine_origin', 'span', 'listGridItemOrigin_text'),
            ('rating_avg', 'span', 'averageRating_average'),
            ('rating_num', 'span', 'averageRating_number'),
            ('price_current', 'span', 'productPrice_price-saleWhole'),
            ('price_prediscount', 'span', 'productPrice_price-regWhole'),
            ('discount_nom', 'span', 'productPrice_savings-amount'),
            ('savings_percent', 'span', 'productPrice_savings-percentage')
        ]
        
        for entry_name, tag, class_name in elements_info:
            elements = soup.body.find_all(tag, class_=class_name)
            if idx < len(elements):
                new_entry[entry_name] = elements[idx].text if tag == 'span' else elements[idx].get('content')
        
        results.append(new_entry)

    # Creating the DataFrame
    df = pd.DataFrame(results)
    return df

# Scrape data from each URL to create a list of individual DFs
dataframes = [scrape_wine_data(url) for url in urls]

dataframes[0].head()

# Concatenate the DataFrames
df = pd.concat(dataframes, ignore_index=True)

df.tail()

# Saving df to .csv
df.to_csv('../data/df.csv', index=False)
