==> code/03_user_recs.ipynb <==
import pandas as pd
import random

---

# Load Wine Products DF:
df = pd.read_csv('../data/df.csv')

## Creating User Preferences DataFrame

Ideally this data would come from Wine.com internally. However, since this data is not publically avalable, a fictitious user preferences dataframe has been created showing user-driven activity and associated ratings per product.

# Function to create df of user interactions
def create_user_interaction_df(df):
    
    # Create a list of unique user_ids (215 unique users)
    unique_users = [str(random.randint(100000, 999999)) for _ in range(215)]

    # Create a list of product_ids from the 'product_id' column in the wine DataFrame df
    product_ids = df['product_id'].tolist()

    # Create a list of event types with specified probabilities (total 850)
    event_types = ["view"] * 400 + ["like"] * 200 + ["rate"] * 100 + ["comment"] * 50 + ["follow"] * 50 + ["bookmark"] * 50
    random.shuffle(event_types)

    # Map event types to event strengths
    event_strength_mapping = {
        'view': 1.0,
        'like': 1.5,
        'rate': 2.0,
        'comment': 3.0,
        'follow': 4.0,
        'bookmark': 4.5
    }
    event_strengths = [event_strength_mapping[event_type] for event_type in event_types]

    # Create the DataFrame
    data = {
        "user_id": random.choices(unique_users, k=850),
        "event_type": event_types,
        "product_id": random.choices(product_ids, k=850),
        "event_strength": event_strengths
    }

    df_user_interactions = pd.DataFrame(data)
    return df_user_interactions

Explain: users can interact with Wine.com in many different ways. Explain types, explain subjective rating of 1.0 to 4.5 for strength. 

# Create user interactions DataFrame
df_user_interactions = create_user_interaction_df(df)

# Explore the DF
df_user_interactions.head()

df_user_interactions.info()

print(f'Of the {df_user_interactions.shape[0]} user interactions, there are {df_user_interactions["user_id"].nunique()} unique users and {df_user_interactions["product_id"].nunique()} unique wine products that were rated.')

# Merge DataFrames with outer join
df_all_interactions = pd.merge(df, df_user_interactions, on='product_id', how='outer')

print(f'The original shape of the dataframe df_all_interactions is {df_all_interactions.shape}.')

# Drop nulls for user_id
df_all_interactions.dropna(subset=['user_id'], inplace=True)

print(f'The shape of the dataframe df_all_interactions after dropping user_id nulls is {df_all_interactions.shape}.')

# Redefine event strength mapping outside of previous function
event_strength_mapping = {
    'view': 1.0,
    'like': 1.5,
    'rate': 2.0,
    'comment': 3.0,
    'follow': 4.0,
    'bookmark': 4.5
}

# Map event types to event strengths
df_all_interactions['event_strength'] = df_all_interactions['event_type'].map(event_strength_mapping)

# Calculate the weighted sum per user
df_all_interactions['user_weighted_sum'] = df_all_interactions.groupby('user_id')['event_strength'].transform('sum')

df_all_interactions.head()

# Saving df to .csv
df_all_interactions.to_csv('../data/df_all_interactions.csv', index=False)
